Raise Errors on Missing Translations
The B2B plugin had a lot of invalid i18n keys that could have been caught in the mega-build, so other plugins wouldn't suffer from test failures related to translation issues in other gems. To remedy this, always raise errors on missing translations, so that even if the plugin doesn't have this turned on, we'll eventually catch it in the megabuild before it's released and breaking tests for our users.